requirements.txt
tdl-xoa-driver>=1.2.0
matplotlib>=3.9.2
pyyaml>=6.0.1
pydantic>=2.0